Defer socket descriptor loop-around problem.
We store virtual socket descriptors as uint16 for some reason. When
running a long experiment, or on large topologies, some nodes descriptor
counter may loop around, causing problems with pipecloud simulator
functionality. For now we lower the starting point of the counter to
delay the problem.

A better solution is to switch to a 32-bit virtual socket descriptor, so
I'm leaving this issue open until that is resolved.

refs #22

+++ src/core/vevent/vepoll.h
@@ -56,7 +56,7 @@ enum vepoll_flags {
 
 typedef struct vepoll_s {
 	in_addr_t addr;
-	uint16_t sockd;
+	int sockd;
 	/* OR'ed with types that are allowed (i.e. readable/writable) */
 	enum vepoll_type available;
 	/* OR'ed with types that vevent is waiting for (i.e. readable/writable) */
